services/backend/pyproject.toml
[tool.poetry]
name = "mini-crm-backend"
version = "0.1.0"
description = "Backend service for the multi-tenant mini CRM"
authors = ["Mini CRM Team <devnull@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "mini_crm", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { version = "^0.29.0", extras = ["standard"] }
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
redis = "^5.0.3"
httpx = "^0.27.0"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-mock = "^3.12.0"
ruff = "^0.4.2"
mypy = "^1.8.0"
types-passlib = "^1.7.7.13"
types-python-jose = "^3.3.4"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
